requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
dj-database-url
dj-rest-auth
Django
django-allauth
django-crispy-forms
django-email
django-environ
django-filter
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
gunicorn
idna
Markdown
MouseInfo
numpy
oauthlib
opencv-contrib-python
packaging
pillow
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python3-openid
pytweening
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3
whitenoise